Move BallistaRegistry to appropriate location
During refactoring `BallistaRegistry` ended up at
very strange location. Proposal moves registry
to to more appropriate location.
