+++ ballista/core/src/serde/scheduler/mod.rs
@@ -15,27 +15,18 @@
 // specific language governing permissions and limitations
 // under the License.
 
-use std::collections::HashSet;
-use std::fmt::Debug;
-use std::{collections::HashMap, fmt, sync::Arc};
-
+use crate::error::BallistaError;
+use crate::registry::BallistaFunctionRegistry;
 use datafusion::arrow::array::{
     ArrayBuilder, StructArray, StructBuilder, UInt64Array, UInt64Builder,
 };
 use datafusion::arrow::datatypes::{DataType, Field};
-use datafusion::common::DataFusionError;
-use datafusion::execution::{FunctionRegistry, SessionState};
-use datafusion::functions::all_default_functions;
-use datafusion::functions_aggregate::all_default_aggregate_functions;
-use datafusion::functions_window::all_default_window_functions;
-use datafusion::logical_expr::planner::ExprPlanner;
-use datafusion::logical_expr::{AggregateUDF, ScalarUDF, WindowUDF};
 use datafusion::physical_plan::ExecutionPlan;
 use datafusion::physical_plan::Partitioning;
 use datafusion::prelude::SessionConfig;
 use serde::Serialize;
-
-use crate::error::BallistaError;
+use std::fmt::Debug;
+use std::{collections::HashMap, fmt, sync::Arc};
 
 pub mod from_proto;
 pub mod to_proto;
@@ -294,90 +285,4 @@ pub struct TaskDefinition {
     pub session_id: String,
     pub session_config: SessionConfig,
     pub function_registry: Arc<BallistaFunctionRegistry>,
-}
-
-#[derive(Debug)]
-pub struct BallistaFunctionRegistry {
-    pub scalar_functions: HashMap<String, Arc<ScalarUDF>>,
-    pub aggregate_functions: HashMap<String, Arc<AggregateUDF>>,
-    pub window_functions: HashMap<String, Arc<WindowUDF>>,
-}
-
-impl Default for BallistaFunctionRegistry {
-    fn default() -> Self {
-        let scalar_functions = all_default_functions()
-            .into_iter()
-            .map(|f| (f.name().to_string(), f))
-            .collect();
-
-        let aggregate_functions = all_default_aggregate_functions()
-            .into_iter()
-            .map(|f| (f.name().to_string(), f))
-            .collect();
-
-        let window_functions = all_default_window_functions()
-            .into_iter()
-            .map(|f| (f.name().to_string(), f))
-            .collect();
-
-        Self {
-            scalar_functions,
-            aggregate_functions,
-            window_functions,
-        }
-    }
-}
-
-impl FunctionRegistry for BallistaFunctionRegistry {
-    fn expr_planners(&self) -> Vec<Arc<dyn ExprPlanner>> {
-        vec![]
-    }
-
-    fn udfs(&self) -> HashSet<String> {
-        self.scalar_functions.keys().cloned().collect()
-    }
-
-    fn udf(&self, name: &str) -> datafusion::common::Result<Arc<ScalarUDF>> {
-        let result = self.scalar_functions.get(name);
-
-        result.cloned().ok_or_else(|| {
-            DataFusionError::Internal(format!(
-                "There is no UDF named \"{name}\" in the TaskContext"
-            ))
-        })
-    }
-
-    fn udaf(&self, name: &str) -> datafusion::common::Result<Arc<AggregateUDF>> {
-        let result = self.aggregate_functions.get(name);
-
-        result.cloned().ok_or_else(|| {
-            DataFusionError::Internal(format!(
-                "There is no UDAF named \"{name}\" in the TaskContext"
-            ))
-        })
-    }
-
-    fn udwf(&self, name: &str) -> datafusion::common::Result<Arc<WindowUDF>> {
-        let result = self.window_functions.get(name);
-
-        result.cloned().ok_or_else(|| {
-            DataFusionError::Internal(format!(
-                "There is no UDWF named \"{name}\" in the TaskContext"
-            ))
-        })
-    }
-}
-
-impl From<&SessionState> for BallistaFunctionRegistry {
-    fn from(state: &SessionState) -> Self {
-        let scalar_functions = state.scalar_functions().clone();
-        let aggregate_functions = state.aggregate_functions().clone();
-        let window_functions = state.window_functions().clone();
-
-        Self {
-            scalar_functions,
-            aggregate_functions,
-            window_functions,
-        }
-    }
 }
